[checkers] new hp_sw checker for StorageWorks hw
This checker is a tri-state : up, down, ghost.

Ghost state is defined as (inq succesful && tur fails)
Up state is defined as (inq succes && tur succesful)
Down state is defined as (inq fails)

multipath modified to print [ghost] as a pstate when appropriate.
multipathd modified to reinstate path in ghost state, as it does with "up".

This final mod was the goal of the patch : indeed the SCSI_START hardware
handler is asked to init a PG only if it has at least one active path.
We previously considered ghost paths as failed, so the daemon proactiveled
deactivated these paths, which kept the hwhandler from being given a
chance to ask for a controler switch by sending a SCSI_START.

Survives my testing here on EVA5000.