[multipathd] uxsock patch
I was playing around with the CLI, and having some issues. Occasionally
commands wouldn't work, which was because multipathd was using string 
functions on strings that weren't NULL terminated. Also, if you just typed
in junk for the name of the map to remove, it would always remove dm-0,
because atoi returned 0.  I fixed those, added some moron-proofing checks,
and made multipathd reply "fail" if the command failed.

Here's the patch

Benjamin Marzinski, Redhat.