[libmultipath] possible daemon crash with pp->wwid == 0x0
It is possible for a path to have its .wwid empty in the daemon pathvec,
causing all sorts of trouble.

In disassemble_map(), update the pp->wwid with mpp->wwid for paths which
may not have been active at the time the getuid callout was run.

Spotted by Edward Goggin, EMC.