[multipath] option to use priorities as weights in the round-robin scheduler
Disabled by default.

Synthax is "rr_weight = priorities", and is valid in the default{},
device{} and multipath{} blocks.

This is an interim solution before a cleaner "least queued scheduler".

To make this complete, the daemon should be able to "message" DM to
change the path weights when their prio changes. Probably not worth the
effort and spend time on another scheduler.

The core of this patch is in the multipath table setup : 

if (mp->rr_weight && pp->priority)
	minio = conf->rr_minio * pp->priority;

The rest is the config switch handling, and printing.