[libmultipath] Fixup hwtable regexp
appearently no-one noticed that we've switched to regular expressions
instead of shell-style fnmatch for the hwtable.
Quite surprisingly, really, that no-one complained; the change has been
done in October 2005.

Anyway, this patch cleans it up. And also splits off the IBM DS6000 and
IBM SVC, as they are separate machines.

Hannes Reinecke, Suse