[libmultipath] Use strmatch to load config file
Currently we're using regexec to load the initial hardware table.
Unfortunately the hardware table has regex as defaults, so it's quite
possible for two different regular expressions do match.
But that's quite undesired as we just want a clean 1:1 copy here.
So we're better using a string match for comparison.

Signed-off-by: Hannes Reinecke <hare@suse.de>