[libmultipath] pathinfo() to invoke the path priority callout
Changed libmultipath/discovery.c to allow pathinfo() to invoke the path
priority callout function of a path with a path state of PATH_DOWN since
some paths (e.g., paths to inactive snapshot LU on CLARiiON) may have
their state set to PATH_DOWN for reasons other than transport
connectivity failure.

Without this change, it is possible to create a situation whereby there
are paths of different priorities in the same path group even though the
path group policy is priority based grouping.  This is certainly the
case for paths to a multipath mapped device for an inactive snapshot
CLARiiON LUs when the paths change to a healthy state as soon as the
snapshot is activated.

Edward Goggin, EMC