[checkers] handling CLARiiON I/O to inactive snapshot logical units
Prevent an I/O hang which can occur with host I/O to an inactive CLARiiON
snapshot logical unit.  While the inactive snapshot LU is presented to a
host, inquiry, TUR, and read capacity commands succeed, but read and write
commands are failed.  Complicating the matter is the fact that read/write
to the active paths fails with a particular sense code/asc/ascq combination
while read/write to the passive paths does not.

Edward Goggin, EMC