[multipath] correctness "multipath -ll" fix
Just a simple change to multipath/main.c:update_paths()
to set the mpp field of each path so that the immediately
following call to pathinfo() will call get_state() on a
path with its  mpp field set.  This enables "multipath -ll"
to utilize the shared context field of the multipath struct.

Edward Goggin, EMC