apply the same treatment to prioritizer than Ben applied to checkers
ie:
o don't so prio shared object if not used
o error if prioritizer lib is not found instead of falling back to default

I also had to move up the multipath_dir initialization in load_config()
because add_prio/add_checker need the var to load the shared object.
Also realize the multpath_dir keyword has to be  set before default:prio
and default:checker in the config file.

A correct solution would be to defer the loads to the end of the config
file processing. Contribution welcome.