add '-r' option to force devmap reload
This takes care of updating the write-protection flag
of the devmap in case the write-prot of the underlying
paths has changed (and the kernel somehow has discovered
so).

Now a RO->RW change scenario for a multipathed device can
be treated with :
1/ echo 1 >/sys/block/XX/device/rescan on all paths to dev
2/ multipath -r dev

Which minimaly fulfills my needs with Symmetrix SRDF spliting
in interim of more collaboration from other subsystems.