Fix for parsing multipath.conf files without spaces before brackets
There's a bug in bug in the multipath.conf code that keeps multipath from
correctly parsing config files where there is no space between a section name
and the opening bracket. For instance

devices {
	device {
		...
	}
}

works but

devices {
	device{
		...
	}
}

doesn't. This patch makes sure that brackets are the recognized as seperate from
the token that they follow, unless they are part of a quoted string.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>