multipath_tools: Fixup IBM Virtual SCSI hwtable entries
Removes a hwtable entry for a IBM Virtual SCSI vendor/device ID that
is no longer going to be released. Adds a new vendor/device ID for
a new IBM Virtual SCSI vendor/device ID. This is needed since path
switching on this device type is expensive, so we don't want to do
round robin. Additionally, the default health checker (direct IO)
does not work for this device type.

Signed-off-by: Brian King <brking@linux.vnet.ibm.com>