multipath-tools: Set the rdac TAS bit if it changeable
This patch sets the rdac TAS bit if it is changeable. This is called only once
per lun during the discovery.
The Task Aborted Status (TAS) bit in Control Mode Page (Page 0x0A) controls
the behavior of aborted tasks. By default, the TAS bit is set to zero. The
rdac checker should issue a mode sense page 0x0A to determine the TAS bit
value and whether or not the TAS bit is changeable. If the TAS bit is set
to 0 and changeable, the rdac checker should issue a mode select page 0x0A
to change TAS to 1 and save it.

Signed-off-by: Babu Moger <babu.moger@lsi.com>
Reviewed-by  : Yanling Qi <yanling.qi@lsi.com>
Reviewed-by  : Somasundaram Krishnasamy <Somasundaram.Krishnasamy@lsi.com>