multipath: don't set queue_if_no_path without multipathd
If multipathd is not running, when all paths to a device have failed, there's
no way for them to automatically get restored.  If the device is set to queue,
whatever is accessing it will hang forever. This can lead to problems if it
happens at boot-up.  This patch unsets queue_if_no_path for all devices created
when multipathd is not running. When multipathd starts, it will automatically
get reset queue_if_no_path to the proper value.  This new behaviour can be
overridden using the new "-q" option to multipath.

This version of the patch contacts multipathd's client socket to tell if it's
running.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>