multipath-tools: Remove bad udev rules
This sample udev rules file contains some rules relating to dmraid
that both should not be there and are broken anyhow.  They should
not be there because firstly, what is a dmraid rule doing in a
kpartx rule file, and secondly, dmraid already activates partitions
itself, so there is no need to run kpartx to do that.  The rule is
broken because it is matching on the DM_UUID starting with "dmraid-",
but this comparison is case sensitive, and it actually starts with
"DMRAID-".

Signed-off-by: Phillip Susi <psusi@cfl.rr.com>