multipath-tools: Adding major:minor to multipathd logging for device discovery and removal
Path name, HCTL and major:minor (for ex: srp and mptsas) could change during subsequent
rediscovery of path and it becomes difficult to associate the paths with dm device using
default verbosity in the syslog (when there are large number of dm device configured).
multipathd logs few messages with path name and others with major:minor. We need to have
association between  major:minor and device name with default verbosity.

Signed-off-by: Vijay Chauhan <vijay.chauhan@netapp.com>