Don't detect prioritizers that don't work
The current method used by detect_prio was selecting the ALUA
prioritizer for devices that didn't have ALUA enabled.  This patch
makes detect_prio go through all the steps to get a priority with the
ALUA prioritizer.  If it is able to successfully get a priority, then
it selects the ALUA proritizer for the device.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>