Only load a multipath device RO if RW load fails with EROFS
If a transient error keeps multipath from being able to reload a
multipath device read/write, it will retry the reload with the
device set to read-only.  This can suddenly turn a read/write
multipath device read-only.  If device-mapper cannot load the
device read/write because a path device can't be opened for writing,
it will return EROFS.  This is the only time it makes sense to
reload the device read-only.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>