multipathd: not skip paths not belonging to a mpp in check_path
commit bf4c80aee4e977e89d6925df569647acd3d5efef
multipathd: add path when transitioned from 'blocked' state
intends to check paths not belonging to a mpp usually because of
they are in 'blocked' state and we cannot get the path uuid.
So these paths ought not to be skipped at the start of the function.
It could be a fault when merge the original patch -
http://www.redhat.com/archives/dm-devel/2013-December/msg00085.html

Signed-off-by: Vaughan Cao <Vaughan.Cao@oracle.com>